+++ internal/examples/l0/broadcast_l0.cpp
@@ -21,7 +21,7 @@
 
 template <class processing_type>
 void user_thread_idx(size_t thread_idx,
-                     std::vector<std::pair<size_t, ccl::device_index_type>> ranked_device_indices,
+                     std::vector<std::pair<int, ccl::device_index_type>> ranked_device_indices,
                      std::shared_ptr<::native::ccl_context> ctx,
                      int total_devices_in_cluster,
                      std::shared_ptr<ccl::kvs_interface> kvs) {
@@ -38,7 +38,7 @@ void user_thread_idx(size_t thread_idx,
     std::vector<processing_type> send_values(COUNT);
     std::iota(send_values.begin(), send_values.end(), 1);
     std::vector<processing_type> recv_values(COUNT, 0);
-    size_t root = 1;
+    int root = 1;
 
     // API
     // Create device communicators
@@ -271,7 +271,7 @@ int main(int argc, char** argv) {
         std::vector<device_type> devices;
         std::tie(thread_id, devices) = *thread_affinity_it;
 
-        std::vector<std::pair<size_t, device_type>> ranked_devices;
+        std::vector<std::pair<int, device_type>> ranked_devices;
         ranked_devices.reserve(devices.size());
         std::transform(devices.begin(),
                        devices.end(),
