+++ tests/unit/l0/kernels/a2a_allreduce_single_device_test.hpp
@@ -28,7 +28,7 @@ TEST_F(a2a_allreduce_single_device_fixture, a2a_allreduce_single_device_mt) {
 
     // test case data
     const size_t buffer_size = 512;
-    const size_t num_thread = 4;
+    const int num_thread = 4;
     constexpr size_t mem_group_count = 3;
     constexpr size_t a2a_mem_group_count = 2;
     constexpr size_t flag_group_count = 3;
